+++ lib/src/model/uniquify_list_model.dart
@@ -15,13 +15,30 @@ class UniquifyListModel<T extends Model> {
   /// Get the list of data for th list view.
   List<T> getItems<T>() => _items ?? [];
 
-  /// Add data to list of items for list view.
-  void addItems(
-    List<T> data, {
+  /// Set the default order by function
+  /// for ordering the list by default.
+  int Function(T, T) get orderBy => _orderBy;
+  int Function(T, T) _orderBy;
+  bool _presortOnItemsAdded = false;
+
+  /// Initialize function.
+  /// If [presortOnItemsAdded] the items
+  /// will not be sorted after it has been
+  /// added. Instead it will be presorted
+  /// on items added to the list.
+  void init({
     int Function(T, T) orderBy,
+    bool presortOnItemsAdded = false,
   }) {
+    _orderBy = orderBy;
+    _presortOnItemsAdded = _presortOnItemsAdded ?? false;
+  }
+
+  /// Add data to list of items for list view.
+  void addItems(List<T> data) {
     if (data == null) return;
-    if (orderBy != null) data.sort(orderBy);
+    if (_presortOnItemsAdded && orderBy != null) 
+      data.sort(orderBy);
 
     for (var each in data) {
       if (each == null) continue;
@@ -33,17 +50,17 @@ class UniquifyListModel<T extends Model> {
       _cache[each.id] = each;
       _items.add(each);
     }
+
+    if (!_presortOnItemsAdded && orderBy != null) 
+      data.sort(orderBy);
   }
 
   /// Replace the whole data with a new list of items
   /// for the list view
-  void replaceItems(
-    List<T> data, {
-    int Function(T, T) orderBy,
-  }) {
+  void replaceItems(List<T> data) {
     if (data == null) return;
     _cache.clear();
     _items.clear();
-    this.addItems(data, orderBy: orderBy);
+    this.addItems(data);
   }
 }